release/resources/backend/pyproject.toml
[project]
name = "ai-meeting-assistant-backend"
version = "0.1.0"
description = "FastAPI sidecar for AI Meeting Assistant"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.1",
  "fastapi>=0.115.11",
  "httpx>=0.28.1",
  "psycopg[binary]>=3.2.6",
  "pydantic-settings>=2.8.1",
  "pypdf>=5.4.0",
  "python-docx>=1.1.2",
  "python-multipart>=0.0.20",
  "sqlalchemy>=2.0.39",
  "uvicorn[standard]>=0.34.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
  "pytest-asyncio>=0.25.3"
]

[build-system]
requires = ["setuptools>=75.8.2"]
build-backend = "setuptools.build_meta"